remove all Set-s method and events
