+++ cloud/blockstore/libs/storage/service/service_actor.h
@@ -452,10 +452,6 @@ private:
         TRequestInfoPtr requestInfo,
         TString input);
 
-    TResultOrError<NActors::IActorPtr> CreateSetTabletBootInfosActor(
-        TRequestInfoPtr requestInfo,
-        TString input);
-
     TResultOrError<NActors::IActorPtr> CreateGetClusterCapacityActor(
         TRequestInfoPtr requestInfo,
         TString input);
