+++ cloud/blockstore/libs/storage/service/ya.make
@@ -7,7 +7,6 @@ SRCS(
     service_actor_actions_backup_path_descriptions.cpp
     service_actor_actions_backup_tablet_boot_infos.cpp
     service_actor_actions_get_tablet_boot_infos.cpp
-    service_actor_actions_set_tablet_boot_infos.cpp
     service_actor_actions_change_disk_device.cpp
     service_actor_actions_change_storage_config.cpp
     service_actor_actions_check_blob.cpp
