+++ cloud/storage/core/libs/hive_proxy/hive_proxy_fallback_actor.cpp
@@ -351,18 +351,4 @@ void THiveProxyFallbackActor::HandleGetTabletBootInfos(
     }
 }
 
-void THiveProxyFallbackActor::HandleSetTabletBootInfos(
-    const TEvHiveProxy::TEvSetTabletBootInfosRequest::TPtr& ev,
-    const NActors::TActorContext& ctx)
-{
-    if (TabletBootInfoBackup) {
-        ctx.Send(ev->Forward(TabletBootInfoBackup));
-    } else {
-        auto response =
-            std::make_unique<TEvHiveProxy::TEvSetTabletBootInfosResponse>(
-                MakeError(E_PRECONDITION_FAILED));
-        NCloud::Reply(ctx, *ev, std::move(response));
-    }
-}
-
 }   // namespace NCloud::NStorage
