+++ cloud/storage/core/libs/hive_proxy/tablet_boot_info_backup.cpp
@@ -342,40 +342,6 @@ void TTabletBootInfoBackup::HandleGetTabletBootInfos(
     NCloud::Reply(ctx, *ev, std::move(response));
 }
 
-void TTabletBootInfoBackup::HandleSetTabletBootInfos(
-    const TEvHiveProxy::TEvSetTabletBootInfosRequest::TPtr& ev,
-    const TActorContext& ctx)
-{
-    using TResponse = TEvHiveProxy::TEvSetTabletBootInfosResponse;
-
-    auto* msg = ev->Get();
-
-    auto& data = *BackupProto.MutableData();
-    data.clear();
-
-    for (auto& entry: msg->TabletBootInfos) {
-        ui64 tabletId = entry.StorageInfoProto.GetTabletID();
-
-        NHiveProxy::NProto::TTabletBootInfo tabletBootInfo;
-        *tabletBootInfo.MutableStorageInfo() =
-            std::move(entry.StorageInfoProto);
-        tabletBootInfo.SetSuggestedGeneration(entry.SuggestedGeneration);
-
-        data[tabletId] = std::move(tabletBootInfo);
-    }
-
-    BackupProtoHasChanged = true;
-    InitialBackupProto.reset();
-
-    NProto::TError error;
-    if (!ReadOnlyMode) {
-        error = Backup(ctx);
-    }
-
-    auto response = std::make_unique<TResponse>(std::move(error));
-    NCloud::Reply(ctx, *ev, std::move(response));
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 
 STFUNC(TTabletBootInfoBackup::StateWork)
@@ -387,7 +353,6 @@ STFUNC(TTabletBootInfoBackup::StateWork)
         HFunc(TEvHiveProxy::TEvBackupTabletBootInfosRequest, HandleBackupTabletBootInfos);
         HFunc(TEvHiveProxy::TEvListTabletBootInfoBackupsRequest, HandleListTabletBootInfoBackups);
         HFunc(TEvHiveProxy::TEvGetTabletBootInfosRequest, HandleGetTabletBootInfos);
-        HFunc(TEvHiveProxy::TEvSetTabletBootInfosRequest, HandleSetTabletBootInfos);
 
         default:
             HandleUnexpectedEvent(ev, LogComponent, __PRETTY_FUNCTION__);
